Refactored working classes into PSR4 namespaces.
