+++ composer.json
@@ -7,5 +7,10 @@
     "require-dev": {
         "phpunit/phpunit": "^5.1",
         "nesbot/carbon": "^1.21"
+    },
+    "autoload": {
+        "psr-4": {
+            "ItsDangerous\\": "src/"
+        }
     }
 }
