+++ itsdangerous.php
@@ -2,7 +2,6 @@
 
 define("EPOCH", 1293840000);
 require 'vendor/autoload.php';
-use Carbon\Carbon;
 
 function is_text_serializer($serializer) {
     return is_string($serializer->dumps(array()));
@@ -16,33 +15,6 @@ function constant_time_compare($val1, $val2) {
         $result |= ord($val1[$i]) ^ ord($val2[$i]);
     return $result == 0;
 }
-
-class BadData extends Exception {}
-class BadPayload extends BadData {
-    public $original_error = null;
-    public function __construct($message, $original_error=null) {
-        parent::__construct($message);
-        $this->original_error = $original_error;
-    }
-}
-class BadSignature extends BadData {
-    public $payload = null;
-    public function __construct($message, $payload=null) {
-        parent::__construct($message);
-        $this->payload = $payload;
-    }
-}
-class BadTimeSignature extends BadData {
-    public $payload = null;
-    public $date_signed = null;
-    public function __construct($message, $payload=null, $date_signed=null) {
-        parent::__construct($message);
-        $this->payload = $payload;
-        $this->date_signed = $date_signed;
-    }
-}
-class SignatureExpired extends BadTimeSignature {}
-
 
 function base64_encode_($data) {
     return rtrim(strtr(base64_encode($data), '+/', '-_'), '=');
@@ -71,336 +43,10 @@ function bytes_to_int($bytes) {
     return $output;
 }
 
-abstract class SigningAlgorithm {
-    abstract public function get_signature($key, $value);
-
-    public function verify_signature($key, $value, $sig) {
-        return constant_time_compare($sig, $this->get_signature($key, $value));
-    }
-}
-
-class NoneAlgorithm extends SigningAlgorithm {
-    public function get_signature($key, $value) {
-        return '';
-    }
-}
-
-class HMACAlgorithm extends SigningAlgorithm {
-
-    public static $default_digest_method = 'sha1';
-
-    private $digest_method;
-
-    public function __construct($digest_method=null) {
-        if (is_null($digest_method)) {
-            $digest_method = self::$default_digest_method;
-        }
-        $this->digest_method = $digest_method;
-    }
-
-    public function get_signature($key, $value) {
-        return hash_hmac($this->digest_method, $value, $key, true);
-    }
-}
 
 
-class Signer {
 
-    public static $default_digest_method = 'sha1';
-    public static $default_key_derivation = 'django-concat';
-
-    protected $secret_key;
-    protected $sep;
-    protected $salt;
-    protected $key_derivation;
-    protected $digest_method;
-    protected $algorithm;
-
-    public function __construct($secret_key, $salt=null, $sep='.', $key_derivation=null, $digest_method=null, $algorithm=null) {
-        $this->secret_key = $secret_key;
-        $this->sep = $sep;
-        $this->salt = is_null($salt) ? 'itsdangerous.Signer' : $salt;
-        $this->key_derivation = is_null($key_derivation) ? self::$default_key_derivation : $key_derivation;
-        $this->digest_method = is_null($digest_method) ? self::$default_digest_method : $digest_method;
-        $this->algorithm = is_null($algorithm) ? new HMACAlgorithm($this->digest_method) : $algorithm;
-    }
-
-    protected function digest($input) {
-        return hash($this->digest_method, $input, true);
-    }
-
-    public function derive_key() {
-    	switch ($this->key_derivation) {
-    	    case 'concat':
-                return $this->digest($this->salt . $this->secret_key);
-            case 'django-concat':
-                return $this->digest($this->salt . 'signer' . $this->secret_key);
-            case 'hmac':
-                return hash_hmac($this->digest_method, $this->salt, $this->secret_key, true);
-            default:
-                throw new Exception("Unknown key derivation method");
-    	}
-    }
-
-    public function get_signature($value) {
-        $key = $this->derive_key();
-        $sig = $this->algorithm->get_signature($key, $value);
-        return base64_encode_($sig);
-    }
-
-    public function sign($value) {
-        return $value . $this->sep . $this->get_signature($value);
-    }
-
-    public function verify_signature($value, $sig) {
-        $key = $this->derive_key();
-        $sig = base64_decode_($sig);
-        return $this->algorithm->verify_signature($key, $value, $sig);
-    }
-
-    public function unsign($signed_value) {
-        if(strpos($signed_value, $this->sep) === false)
-            throw new BadSignature("No \"{$this->sep}\" found in value");
-        $exploded = explode($this->sep, $signed_value);
-        $sig = array_pop($exploded);
-        $value = implode($this->sep, $exploded);
-        if($this->verify_signature($value, $sig))
-            return $value;
-        throw new BadSignature("Signature \"{$sig}\" does not match", $value);
-    }
-
-    public function validate($signed_value) {
-        try {
-            $this->unsign($signed_value);
-            return true;
-        } catch(BadSignature $ex) {
-            return false;
-        }
-    }
-
-}
-
-class TimestampSigner extends Signer {
-
-    public function get_timestamp() {
-        return Carbon::now()->timestamp - EPOCH;
-    }
-
-    public function timestamp_to_datetime($ts) {
-        return DateTime::createFromFormat("U", $ts + EPOCH, new DateTimeZone("UTC"));
-    }
-
-    public function sign($value) {
-        $timestamp = base64_encode_(int_to_bytes($this->get_timestamp()));
-        $value = $value . $this->sep . $timestamp;
-        return $value . $this->sep . $this->get_signature($value);
-    }
-
-    public function unsign($value, $max_age=null, $return_timestamp=false) {
-
-        try {
-            $result = parent::unsign($value);
-            $sig_err = null;
-        } catch (BadSignature $ex) {
-            $sig_err = $ex;
-            $result = $ex->payload;
-        }
-
-        if(strpos($result, $this->sep) === false) {
-            if (!is_null($sig_err)) {
-                throw $sig_err;
-            }
-            throw new BadTimeSignature("timestamp missing", $result);
-        }
-
-        $exploded = explode($this->sep, $result);
-        $timestamp = array_pop($exploded);
-        $value = implode($this->sep, $exploded);
-
-        try {
-            $timestamp = bytes_to_int(base64_decode_($timestamp));
-        // @codeCoverageIgnoreStart
-        // not sure if this try catch is even needed - the builtins
-        // involved don't throw like their counter parts in python
-        } catch (Exception $ex) {
-            $timestamp = null;
-        }
-        // @codeCoverageIgnoreEnd
-
-        # Signature is *not* okay.  Raise a proper error now that we have
-        # split the value and the timestamp.
-        if (!is_null($sig_err))
-            throw new BadTimeSignature((string) $sig_err, $value, $timestamp);
-
-        # Signature was okay but the timestamp is actually not there or
-        # malformed.  Should not happen, but well.  We handle it nonetheless
-        if (is_null($timestamp))
-            throw new BadTimeSignature('Malformed timestamp', $value);
-
-        if(!is_null($max_age)) {
-            $age = $this->get_timestamp() - $timestamp;
-            if($age > $max_age)
-                throw new SignatureExpired(
-                    "Signature age $age > $max_age seconds",
-                    $value,
-                    $this->timestamp_to_datetime($timestamp));
-        }
-
-        if($return_timestamp)
-            return array($value, $this->timestamp_to_datetime($timestamp));
-        return $value;
-    }
-
-    public function validate($signed_value, $max_age=null) {
-        try {
-            $this->unsign($signed_value, $max_age);
-            return true;
-        } catch(Exception $ex) {
-            return false;
-        }
-    }
-
-}
-
-class simplejson {
-
-    public function loads($input) {return json_decode($input, true);}
-    public function dumps($input) {return json_encode($input);}
-
-}
-
-class Serializer {
-
-    public static $default_signer = 'Signer';
-    public static function default_serializer() {return new simplejson();}
-
-    protected $secret_key;
-    protected $salt;
-    protected $serializer;
-    protected $is_text_serializer;
-
-    public function __construct($secret_key, $salt="itsdangerous", $serializer=null, $signer=null) {
-        $this->secret_key = $secret_key;
-        $this->salt = $salt;
-        if(is_null($serializer))
-            $serializer = self::default_serializer();
-        $this->serializer = $serializer;
-        $this->is_text_serializer = is_text_serializer($this->serializer);
-        if (is_null($signer))
-            $signer = self::$default_signer;
-        $this->signer = $signer;
-    }
-
-    public function load_payload($payload, $serializer=null) {
-        if (is_null($serializer)) {
-            $serializer = $this->serializer;
-            $is_text = $this->is_text_serializer;
-        } else {
-            // @codeCoverageIgnoreStart
-            // $is_text is never used for that matter, is_text_serializer
-            // doesn't seemed to be used outside of this
-            $is_text = is_text_serializer($serializer);
-            // @codeCoverageIgnoreEnd
-        }
-        try {
-            return $serializer->loads($payload);
-        } catch (Exception $ex) {
-            throw new BadPayload(
-                "Could not load the payload because an exception occurred " +
-                "on unserializing the data.", $ex);
-        }
-    }
-
-    public function dump_payload($obj) {
-        return $this->serializer->dumps($obj);
-    }
-
-    public function make_signer($salt=null) {
-        if (is_null($salt)) {
-            $salt = $this->salt;
-        }
-        $signer = $this->signer;
-        return new $signer($this->secret_key, $salt);
-    }
-
-    public function dumps($obj, $salt=null) {
-        return $this->make_signer($salt)->sign($this->dump_payload($obj));
-    }
-
-    public function dump($obj, $f, $salt=null) {
-        fwrite($f, $this->dumps($obj, $salt));
-    }
-
-    public function loads($s, $salt=null) {
-        return $this->load_payload($this->make_signer($salt)->unsign($s));
-    }
-
-    // @codeCoverageIgnoreStart
-    // lolwut - filesize of a file pointer or fread on a filename?
-    public function load($f, $salt=null) {
-        return $this->loads(fread($f, filesize($f)), $salt);
-    }
-    // @codeCoverageIgnoreEnd
-
-    public function loads_unsafe($s, $salt=null) {
-        try {
-            return array(true, $this->loads($s, $salt));
-        } catch (Exception $ex) {
-            if (empty($ex->payload)) {
-                return array(false, null);
-            }
-            try {
-                return array(false, $this->load_payload($ex->payload));
-            // @codeCoverageIgnoreStart
-            // I'm confused by the circular logic here - we just caught
-            // and exception from loads (maybe with a salt), and we got
-            // a payload in the exception, so we're going to call something
-            // that doesn't do much but call loads again and maybe catch
-            // an exception which would mean the data's garbage?
-            } catch (BadPayload $ex) {
-                return array(false, null);
-            }
-            // @codeCoverageIgnoreEnd
-        }
-    }
-
-    // @codeCoverageIgnoreStart
-    // broken because same reason load is broken
-    public function load_unsafe($f, $salt=null) {
-        return $this->loads_unsafe(fread($f, filesize($f)), $salt);
-    }
-    // @codeCoverageIgnoreEnd
-
-    /*
-    public function _urlsafe_load_payload($payload){
-        $decompress = false;
-        if ($payload[0] == '.'){
-            $payload = substr($payload, 1);
-            $decompress = true;
-        }
-        $json = base64_decode_($payload);
-        if ($decompress){
-            $json = gzuncompress($json);
-
-        }
-        return $json;
-    }
-
-    public function _urlsafe_dump_payload($json){
-        $is_compressed = false;
-        $compressed = gzcompress($json);
-        if (strlen($compressed) < strlen($json) - 1){
-            $json = $compressed;
-            $is_compressed = true;
-        }
-        $base64d = base64_encode_($json);
-        if ($is_compressed){
-            $base64d = '.' . $base64d;
-        }
-        return $base64d;
-    }
-    */
-}
+use ItsDangerous\Signer\Serializer;
 
 /**
  * @codeCoverageIgnore
