+++ phpunit.xml
@@ -6,7 +6,7 @@
     </testsuites>
 	<filter>
 		<whitelist processUncoveredFilesFromWhitelist="true">
-		<!--<directory suffix=".php">/path/to/files</directory>-->
+		<directory suffix=".php">src</directory>-->
 		<file>itsdangerous.php</file>
 		</whitelist>
 	</filter>
