+++ src/BadData/BadData.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace ItsDangerous\BadData;
+use Exception;
+
+class BadData extends Exception {}
