+++ src/BadData/SignatureExpired.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace ItsDangerous\BadData;
+
+class SignatureExpired extends BadTimeSignature {}
