+++ src/Signer/NoneAlgorithm.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ItsDangerous\Signer;
+
+class NoneAlgorithm extends SigningAlgorithm {
+    public function get_signature($key, $value) {
+        return '';
+    }
+}
