+++ src/Signer/Signer.php
@@ -1,0 +1,82 @@
+<?php
+
+namespace ItsDangerous\Signer;
+
+use ItsDangerous\Signer\HMACAlgorithm;
+use ItsDangerous\BadData\BadSignature;
+
+class Signer {
+
+    public static $default_digest_method = 'sha1';
+    public static $default_key_derivation = 'django-concat';
+
+    protected $secret_key;
+    protected $sep;
+    protected $salt;
+    protected $key_derivation;
+    protected $digest_method;
+    protected $algorithm;
+
+    public function __construct($secret_key, $salt=null, $sep='.', $key_derivation=null, $digest_method=null, $algorithm=null) {
+        $this->secret_key = $secret_key;
+        $this->sep = $sep;
+        $this->salt = is_null($salt) ? 'itsdangerous.Signer' : $salt;
+        $this->key_derivation = is_null($key_derivation) ? self::$default_key_derivation : $key_derivation;
+        $this->digest_method = is_null($digest_method) ? self::$default_digest_method : $digest_method;
+        $this->algorithm = is_null($algorithm) ? new HMACAlgorithm($this->digest_method) : $algorithm;
+    }
+
+    protected function digest($input) {
+        return hash($this->digest_method, $input, true);
+    }
+
+    public function derive_key() {
+    	switch ($this->key_derivation) {
+    	    case 'concat':
+                return $this->digest($this->salt . $this->secret_key);
+            case 'django-concat':
+                return $this->digest($this->salt . 'signer' . $this->secret_key);
+            case 'hmac':
+                return hash_hmac($this->digest_method, $this->salt, $this->secret_key, true);
+            default:
+                throw new \Exception("Unknown key derivation method");
+    	}
+    }
+
+    public function get_signature($value) {
+        $key = $this->derive_key();
+        $sig = $this->algorithm->get_signature($key, $value);
+        return base64_encode_($sig);
+    }
+
+    public function sign($value) {
+        return $value . $this->sep . $this->get_signature($value);
+    }
+
+    public function verify_signature($value, $sig) {
+        $key = $this->derive_key();
+        $sig = base64_decode_($sig);
+        return $this->algorithm->verify_signature($key, $value, $sig);
+    }
+
+    public function unsign($signed_value) {
+        if(strpos($signed_value, $this->sep) === false)
+            throw new BadSignature("No \"{$this->sep}\" found in value");
+        $exploded = explode($this->sep, $signed_value);
+        $sig = array_pop($exploded);
+        $value = implode($this->sep, $exploded);
+        if($this->verify_signature($value, $sig))
+            return $value;
+        throw new BadSignature("Signature \"{$sig}\" does not match", $value);
+    }
+
+    public function validate($signed_value) {
+        try {
+            $this->unsign($signed_value);
+            return true;
+        } catch(BadSignature $ex) {
+            return false;
+        }
+    }
+
+}
